Time: 0 ms (100.00%) | Memory: 8.7 MB (41.27%) - LeetSync

+++ 8-string-to-integer-atoi/string-to-integer-atoi.c
@@ -1,6 +1,6 @@
 int myAtoi(char* s) {
     int i=0;
-    long long ans=0;
+    long ans=0;
     int sign=1;
     while(i<strlen(s) && s[i]==' '){
         i++;
